Update issue templates
I have a new system that I am working on for issue submission and these templates make up the bulk of it. I believe that what I am making is more tailored to our specific needs than what Github provides and I think this is the reason they have allowed us to edit it. Details to follow on Hangouts :)