Fixed input buffer in lobby
Added functions `scr_array_sum`, and `scr_positive`. `scr_array_sum` is used to sum up all the values of the inputs for the gamepad, if it is -5, -1 being released, then the game knows nothing was pressed. `scr_positive` is used in the menu to simply the math, making anything pressed a 1 and everything else a 0. Broke menus buttons past lobby.